requirements.txt
asgiref==3.2.10
Django==3.1.1
django-crispy-forms==1.9.2
djangorestframework==3.11.1
joblib==0.16.0
numpy==1.19.2
pandas==1.1.2
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.1.0
